+++ backend/integration_tests/src/disappearing_message_tests.rs
@@ -1,3 +1,4 @@
+use crate::client::community::STABLE_MEMORY_MAP_MEMORY_ID;
 use crate::client::group::CHAT_EVENTS_MEMORY_ID;
 use crate::env::ENV;
 use crate::stable_memory::get_stable_memory_map;
@@ -226,6 +227,8 @@ fn stable_memory_garbage_collected_after_messages_disappear() {
         },
     );
 
+    let initial_stable_memory_map_keys = get_stable_memory_map(env, group_id, STABLE_MEMORY_MAP_MEMORY_ID).len();
+
     for _ in 0..5 {
         let result = client::group::happy_path::send_text_message(env, &user, group_id, None, random_string(), None);
 
@@ -241,7 +244,10 @@ fn stable_memory_garbage_collected_after_messages_disappear() {
         }
     }
 
-    assert_eq!(get_stable_memory_map(env, group_id, CHAT_EVENTS_MEMORY_ID).len(), 33);
+    assert_eq!(
+        get_stable_memory_map(env, group_id, CHAT_EVENTS_MEMORY_ID).len(),
+        initial_stable_memory_map_keys + 30
+    );
 
     // Tick once to expire the messages
     env.advance_time(Duration::from_secs(2));
@@ -251,5 +257,8 @@ fn stable_memory_garbage_collected_after_messages_disappear() {
     env.advance_time(Duration::from_secs(60));
     env.tick();
 
-    assert_eq!(get_stable_memory_map(env, group_id, CHAT_EVENTS_MEMORY_ID).len(), 3);
+    assert_eq!(
+        get_stable_memory_map(env, group_id, CHAT_EVENTS_MEMORY_ID).len(),
+        initial_stable_memory_map_keys
+    );
 }