+++ backend/libraries/principal_to_user_id_map/src/lib.rs
@@ -1,0 +1,69 @@
+use ic_principal::Principal;
+use serde::de::SeqAccess;
+use serde::{Deserialize, Deserializer, Serialize};
+use stable_memory_map::{with_map, with_map_mut, KeyPrefix, PrincipalToUserIdKeyPrefix};
+use std::fmt::Formatter;
+use types::UserId;
+
+#[derive(Serialize, Deserialize, Default)]
+pub struct PrincipalToUserIdMap {
+    prefix: PrincipalToUserIdKeyPrefix,
+    count: u32,
+}
+
+impl PrincipalToUserIdMap {
+    pub fn get(&self, principal: &Principal) -> Option<UserId> {
+        with_map(|m| m.get(self.prefix.create_key(principal)).map(bytes_to_user_id))
+    }
+
+    pub fn insert(&mut self, principal: Principal, user_id: UserId) {
+        if with_map_mut(|m| m.insert(self.prefix.create_key(&principal), user_id.as_slice().to_vec())).is_none() {
+            self.count += 1;
+        }
+    }
+
+    pub fn remove(&mut self, principal: &Principal) -> Option<UserId> {
+        let bytes = with_map_mut(|m| m.remove(self.prefix.create_key(principal)))?;
+        self.count = self.count.saturating_sub(1);
+        Some(bytes_to_user_id(bytes))
+    }
+
+    pub fn len(&self) -> u32 {
+        self.count
+    }
+
+    pub fn is_empty(&self) -> bool {
+        self.count == 0
+    }
+}
+
+pub fn deserialize_principal_to_user_id_map_from_heap<'de, D: Deserializer<'de>>(
+    d: D,
+) -> Result<PrincipalToUserIdMap, D::Error> {
+    d.deserialize_seq(Visitor)
+}
+
+struct Visitor;
+
+impl<'a> serde::de::Visitor<'a> for Visitor {
+    type Value = PrincipalToUserIdMap;
+
+    fn expecting(&self, formatter: &mut Formatter) -> std::fmt::Result {
+        formatter.write_str("a sequence of (Principal, UserId)")
+    }
+
+    fn visit_seq<A>(self, mut seq: A) -> Result<Self::Value, A::Error>
+    where
+        A: SeqAccess<'a>,
+    {
+        let mut map = PrincipalToUserIdMap::default();
+        while let Some((principal, user_id)) = seq.next_element()? {
+            map.insert(principal, user_id);
+        }
+        Ok(map)
+    }
+}
+
+fn bytes_to_user_id(bytes: Vec<u8>) -> UserId {
+    UserId::from(Principal::from_slice(&bytes))
+}